resolved bug in the condition

+++ tests/integrity_tests/compileNode.h
@@ -13,5 +13,5 @@ void compile_states(xmlNode * a_node,Rules *rules);
  void compile_xagent(xmlNode * a_node,Rules *rules);
 
 char *compile_name(xmlNode * a_node);
-float compile_var(xmlNode * a_node);
+double compile_var(xmlNode * a_node);
 #endif
